+++ src/helpers/time_helpers.ts
@@ -1,6 +1,7 @@
 /*
  * Helpers for time-related operations
  */
+import Storage from "../Storage";
 import { NULL, quote} from "./string_helpers";
 
 export const SECONDS_IN_MINUTE = 60;
@@ -75,4 +76,11 @@ export const timeString = (_timestamp: Date | string | null, locale?: string): s
 export function nowString(): string {
     const now = new Date();
     return `${now.toLocaleDateString()} ${now.toLocaleTimeString().split(".")[0]}`;
+};
+
+
+// Return the oldest timestamp we should feed timeline toots until
+export function timelineCutoffAt(): Date {
+    const timelineLookBackMS = Storage.getConfig().maxTimelineHoursToFetch * 3600 * 1000;
+    return new Date(Date.now() - timelineLookBackMS);
 };